backup_before_refactor/project/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.62.0
albumentations>=1.3.0
opencv-python>=4.5.0
timm>=0.9.0
Pillow>=8.3.0
grad-cam>=1.4.0
efficientnet-pytorch>=0.7.1
tensorboard>=2.11.0 